add the option to raise execptions on planning or execution failure
also refactor some classes so this file doesn't get too long
