+++ arm_robots/src/arm_robots/robot.py
@@ -1,5 +1,4 @@
 #! /usr/bin/env python
-from dataclasses import dataclass
 from typing import List, Union, Tuple, Callable, Optional, Dict
 
 import numpy as np
@@ -11,72 +10,20 @@ import rospy
 from actionlib import SimpleActionClient
 from arc_utilities.conversions import convert_to_pose_msg
 from arm_robots.base_robot import DualArmRobot
-from arm_robots.robot_utils import make_follow_joint_trajectory_goal
+from arm_robots.robot_utils import make_follow_joint_trajectory_goal, PlanningResult, PlanningAndExecutionResult, \
+    ExecutionResult
 from control_msgs.msg import FollowJointTrajectoryAction, FollowJointTrajectoryFeedback, FollowJointTrajectoryResult, \
     FollowJointTrajectoryGoal
 from geometry_msgs.msg import Point, Pose, Quaternion
-from moveit_msgs.msg import RobotTrajectory, MoveItErrorCodes
+from moveit_msgs.msg import RobotTrajectory
 from rosgraph.names import ns_join
 from rospy import logfatal
-from rospy.logger_level_service_caller import LoggerLevelServiceCaller
 from trajectory_msgs.msg import JointTrajectory, JointTrajectoryPoint
 
 STORED_ORIENTATION = None
 
 store_error_msg = ("No stored tool orientations! "
                    "You have to call store_tool_orientations or store_current_tool_orientations first")
-
-
-@dataclass
-class ExecutionResult:
-    trajectory: Optional[JointTrajectory]
-    execution_result: Optional[FollowJointTrajectoryResult]
-    action_client_state: int
-    success: bool
-
-
-class PlanningResult:
-
-    def __init__(self,
-                 move_group_plan_tuple: Optional[Tuple[bool, RobotTrajectory, float, MoveItErrorCodes]] = None,
-                 success: Optional[bool] = None,
-                 plan: Optional[RobotTrajectory] = None,
-                 planning_time: Optional[float] = None,
-                 planning_error_code: Optional[int] = None,
-                 ):
-        if move_group_plan_tuple is not None:
-            self.success = move_group_plan_tuple[0]
-            self.plan = move_group_plan_tuple[1]
-            self.planning_time = move_group_plan_tuple[2]
-            self.planning_error_code = move_group_plan_tuple[3]
-        else:
-            self.success = success
-            self.plan = plan
-            self.planning_time = planning_time
-            self.planning_error_code = planning_error_code
-
-
-class PlanningAndExecutionResult:
-    def __init__(self, planning_result: PlanningResult, execution_result: ExecutionResult):
-        self.planning_result = planning_result
-        self.execution_result = execution_result
-        self.success = planning_result.success or execution_result.success
-
-
-def set_move_group_log_level(event):
-    # Up the logging level for MoveGroupInterface because it's annoying
-    log_level = LoggerLevelServiceCaller()
-    node_name = "cpp_" + rospy.get_name().strip("/")
-    logger_name = "ros.moveit_ros_planning_interface.move_group_interface"
-    node_names = log_level.get_node_names()
-    if node_name in node_names:
-        loggers = log_level.get_loggers(node_name)
-        if logger_name in loggers:
-            try:
-                success = log_level.send_logger_change_message(node_name, logger_name, "WARN")
-                rospy.logdebug_once(f"status of changing move_group_interface logger level: {success}")
-            except Exception:
-                pass
 
 
 class MoveitEnabledRobot(DualArmRobot):
@@ -86,10 +33,12 @@ class MoveitEnabledRobot(DualArmRobot):
                  arms_controller_name: str,
                  execute: bool = True,
                  block: bool = True,
+                 raise_on_failure: bool = False,
                  force_trigger: float = 9.0):
         super().__init__(robot_namespace)
         self.max_velocity_scale_factor = 0.1
         self.stored_tool_orientations = None
+        self.raise_on_failure = raise_on_failure
         self.execute = execute
         self.block = block
         self.force_trigger = force_trigger
@@ -104,9 +53,6 @@ class MoveitEnabledRobot(DualArmRobot):
         self.jacobian_follower = None
 
         self.feedback_callbacks = []
-
-        # Up the logging level for MoveGroupInterface because it's annoying
-        # rospy.Timer(rospy.Duration(2), set_move_group_log_level)
 
     def connect(self):
         # TODO: bad api? raii? this class isn't fully usable by the time it's constructor finishes, that's bad.
@@ -148,7 +94,11 @@ class MoveitEnabledRobot(DualArmRobot):
         move_group = self.get_move_group_commander(group_name)
         move_group.set_end_effector_link(ee_link_name)
         move_group.set_position_target(list(target_position))
+
         planning_result = PlanningResult(move_group.plan())
+        if self.raise_on_failure and not planning_result.success:
+            raise RuntimeError(f"Plan to position failed {planning_result.planning_error_code}")
+
         execution_result = self.follow_arms_joint_trajectory(planning_result.plan.joint_trajectory)
         return PlanningAndExecutionResult(planning_result, execution_result)
 
@@ -170,10 +120,12 @@ class MoveitEnabledRobot(DualArmRobot):
             waypoint_pose.position.y = target_position[1]
             waypoint_pose.position.z = target_position[2]
         waypoints = [waypoint_pose]
+
         plan, fraction = move_group.compute_cartesian_path(waypoints=waypoints, eef_step=step_size, jump_threshold=0.0)
-        if fraction != 1.0:
-            raise RuntimeError(f"Cartesian path is only {fraction * 100}% complete")
         planning_result = PlanningResult(success=(fraction == 1.0), plan=plan)
+        if self.raise_on_failure and not planning_result.success:
+            raise RuntimeError(f"Cartesian path is only {fraction * 100}% complete")
+
         execution_result = self.follow_arms_joint_trajectory(plan.joint_trajectory)
         return PlanningAndExecutionResult(planning_result, execution_result)
 
@@ -188,6 +140,9 @@ class MoveitEnabledRobot(DualArmRobot):
         move_group.set_goal_orientation_tolerance(0.02)
 
         planning_result = PlanningResult(move_group.plan())
+        if self.raise_on_failure and not planning_result.success:
+            raise RuntimeError(f"Plan to pose failed {planning_result.planning_error_code}")
+
         execution_result = self.follow_arms_joint_trajectory(planning_result.plan.joint_trajectory)
         return PlanningAndExecutionResult(planning_result, execution_result)
 
@@ -217,6 +172,8 @@ class MoveitEnabledRobot(DualArmRobot):
         else:
             move_group.set_joint_value_target(joint_config)
         planning_result = PlanningResult(move_group.plan())
+        if self.raise_on_failure and not planning_result.success:
+            raise RuntimeError(f"Plan to position failed {planning_result.planning_error_code}")
         execution_result = self.follow_arms_joint_trajectory(planning_result.plan.joint_trajectory)
         return PlanningAndExecutionResult(planning_result, execution_result)
 
@@ -247,10 +204,14 @@ class MoveitEnabledRobot(DualArmRobot):
                 if stop_condition is not None and stop_condition(feedback):
                     client.cancel_all_goals()
 
+            # NOTE: this is where execution is actually requested in the form of a joint trajectory
             client.send_goal(goal, feedback_cb=_feedback_cb)
             if self.block:
                 client.wait_for_result()
                 result = client.get_result()
+            if self.raise_on_failure and result != FollowJointTrajectoryResult.SUCCESSFUL:
+                raise RuntimeError(f"Follow Joint Trajectory Failed: ({result.error_code}) {result.error_string}")
+
         return ExecutionResult(trajectory=trajectory,
                                execution_result=result,
                                action_client_state=client.get_state(),
@@ -322,7 +283,11 @@ class MoveitEnabledRobot(DualArmRobot):
             max_velocity_scaling_factor=vel_scaling,
             max_acceleration_scaling_factor=0.1,
         )
-        planning_result = PlanningResult(success=True, plan=robot_trajectory_msg)
+        planning_success = True
+        planning_result = PlanningResult(success=planning_success, plan=robot_trajectory_msg)
+        if self.raise_on_failure and not planning_success:
+            raise RuntimeError(f"Jacobian planning failed")
+
         execution_result = self.follow_arms_joint_trajectory(robot_trajectory_msg.joint_trajectory,
                                                              stop_condition=stop_condition)
         return PlanningAndExecutionResult(planning_result, execution_result)
